add enable/disable for project retention

+++ src/main/java/org/dependencytrack/tasks/maintenance/ProjectMaintenanceTask.java
@@ -24,6 +24,7 @@ import alpine.event.framework.Subscriber;
 import org.dependencytrack.event.maintenance.ProjectMaintenanceEvent;
 import org.dependencytrack.persistence.jdbi.ConfigPropertyDao;
 import org.dependencytrack.persistence.jdbi.ProjectDao;
+import org.jdbi.v3.core.Handle;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -32,9 +33,11 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static net.javacrumbs.shedlock.core.LockAssert.assertLocked;
 import static org.dependencytrack.model.ConfigPropertyConstants.MAINTENANCE_PROJECTS_RETENTION_DAYS;
+import static org.dependencytrack.model.ConfigPropertyConstants.MAINTENANCE_PROJECTS_RETENTION_ENABLE;
 import static org.dependencytrack.model.ConfigPropertyConstants.MAINTENANCE_PROJECTS_RETENTION_TYPE;
 import static org.dependencytrack.model.ConfigPropertyConstants.MAINTENANCE_PROJECTS_RETENTION_VERSIONS;
 import static org.dependencytrack.persistence.jdbi.JdbiFactory.inJdbiTransaction;
+import static org.dependencytrack.persistence.jdbi.JdbiFactory.openJdbiHandle;
 import static org.dependencytrack.persistence.jdbi.JdbiFactory.withJdbiHandle;
 import static org.dependencytrack.util.LockProvider.executeWithLock;
 import static org.dependencytrack.util.TaskUtil.getLockConfigForTask;
@@ -73,43 +76,51 @@ public class ProjectMaintenanceTask implements Subscriber {
 
     private Statistics informLocked() {
         assertLocked();
-
-        final String retentionType = withJdbiHandle(handle ->
-                handle.attach(ConfigPropertyDao.class).getValue(MAINTENANCE_PROJECTS_RETENTION_TYPE, String.class));
-        int batchSize = 100;
         AtomicInteger numDeletedTotal = new AtomicInteger(0);
 
-        if (retentionType.equals("AGE")) {
-            final int retentionDays = withJdbiHandle(handle ->
-                    handle.attach(ConfigPropertyDao.class).getValue(MAINTENANCE_PROJECTS_RETENTION_DAYS, Integer.class));
-            final Duration retentionDuration = Duration.ofDays(retentionDays);
-            Instant retentionCutOff = Instant.now().minus(retentionDuration);
-            Integer numDeletedLastBatch = null;
-            while (numDeletedLastBatch == null || numDeletedLastBatch > 0) {
-                numDeletedLastBatch = withJdbiHandle(
-                        batchHandle -> {
-                            final var projectDao = batchHandle.attach(ProjectDao.class);
-                            return projectDao.deleteInactiveProjectsForRetentionDuration(retentionCutOff, batchSize);
-                        });
-                numDeletedTotal.addAndGet(numDeletedLastBatch);
+        final Handle jdbiHandle = openJdbiHandle();
+        final var configPropertyDao = jdbiHandle.attach(ConfigPropertyDao.class);
+        if (configPropertyDao.getValue(MAINTENANCE_PROJECTS_RETENTION_ENABLE, Boolean.class)) {
+
+            final String retentionType = configPropertyDao.getValue(MAINTENANCE_PROJECTS_RETENTION_TYPE, String.class);
+
+            int batchSize = 100;
+            if (retentionType.equals("AGE")) {
+
+                final int retentionDays = configPropertyDao.getValue(MAINTENANCE_PROJECTS_RETENTION_DAYS, Integer.class);
+                jdbiHandle.close();
+                final Duration retentionDuration = Duration.ofDays(retentionDays);
+                Instant retentionCutOff = Instant.now().minus(retentionDuration);
+                Integer numDeletedLastBatch = null;
+                while (numDeletedLastBatch == null || numDeletedLastBatch > 0) {
+                    numDeletedLastBatch = withJdbiHandle(
+                            batchHandle -> {
+                                final var projectDao = batchHandle.attach(ProjectDao.class);
+                                return projectDao.deleteInactiveProjectsForRetentionDuration(retentionCutOff, batchSize);
+                            });
+                    numDeletedTotal.addAndGet(numDeletedLastBatch);
+                }
+            } else {
+                final int versionCountThreshold = configPropertyDao.getValue(MAINTENANCE_PROJECTS_RETENTION_VERSIONS, Integer.class);
+                jdbiHandle.close();
+                Integer projectLastBatch = null;
+                while (projectLastBatch == null || projectLastBatch > 0) {
+                    projectLastBatch = inJdbiTransaction(
+                            batchHandle -> {
+                                final var projectDao = batchHandle.attach(ProjectDao.class);
+                                List<String> projectBatch = projectDao.getDistinctProjects(versionCountThreshold, batchSize);
+                                for (var projectName : projectBatch) {
+                                    numDeletedTotal.addAndGet(projectDao.retainLastXInactiveProjects(projectName, versionCountThreshold));
+                                }
+                                return projectBatch.size();
+                            });
+                }
             }
         } else {
-            final int versionCountThreshold = withJdbiHandle(handle ->
-                    handle.attach(ConfigPropertyDao.class).getValue(MAINTENANCE_PROJECTS_RETENTION_VERSIONS, Integer.class));
-            Integer projectLastBatch = null;
-            while (projectLastBatch == null || projectLastBatch > 0) {
-                projectLastBatch = inJdbiTransaction(
-                        batchHandle -> {
-                            final var projectDao = batchHandle.attach(ProjectDao.class);
-                            List<String> projectBatch = projectDao.getDistinctProjects(versionCountThreshold, batchSize);
-                            for (var projectName : projectBatch) {
-                                numDeletedTotal.addAndGet(projectDao.retainLastXInactiveProjects(projectName, versionCountThreshold));
-                            }
-                            return projectBatch.size();
-                        });
-            }
+            LOGGER.info("Not deleting inactive projects because %s:%s is disabled".formatted(
+                    MAINTENANCE_PROJECTS_RETENTION_ENABLE.getGroupName(),
+                    MAINTENANCE_PROJECTS_RETENTION_ENABLE.getPropertyName()));
         }
         return new Statistics(numDeletedTotal.get());
     }
-
 }
